Added cached container utility.
This is a non-owning container that can be used for caching
where the key is used to check if an item is still available.
Internally only a weak_ptr is kept. As long as something in
the application is still using a ref, then this cached container
can return a shared_ptr to it based on the given key.